+++ AngularClient/src/main/project/project.module.ts
@@ -6,16 +6,15 @@ import { SharedModule } from "../shared/shared.module";
 import { AuthGuard } from "../core/auth-guard.service";
 import { DynamicTitleResolve } from "../core/dynamic-title-resolve.service";
 import { CanDeactivateGuard } from "../core/can-deactivate-guard.service";
-import { NgChartModule } from "../shared/ng-chart/ng-chart.module";
 import { ElementManagerComponent } from "./element-manager.component";
 import { ProjectManagerComponent } from "./project-manager.component";
 import { ProjectViewerComponent } from "./project-viewer.component";
 import { SymbolicPipe } from "./symbolic.pipe";
 
 const projectRoutes: Routes = [
-    { path: ":username/new", component: ProjectManagerComponent, canDeactivate: [CanDeactivateGuard], resolve: { title: DynamicTitleResolve } },
-    { path: ":username/:projectKey/edit", component: ProjectManagerComponent, canActivate: [AuthGuard], canDeactivate: [CanDeactivateGuard], resolve: { title: DynamicTitleResolve } },
-    { path: ":username/:projectKey", component: ProjectViewerComponent, resolve: { title: DynamicTitleResolve } }
+    { path: "projects/new", component: ProjectManagerComponent, canDeactivate: [CanDeactivateGuard], resolve: { title: DynamicTitleResolve } },
+    { path: "projects/:project-id/edit", component: ProjectManagerComponent, canActivate: [AuthGuard], canDeactivate: [CanDeactivateGuard], resolve: { title: DynamicTitleResolve } },
+    { path: "projects/:project-id", component: ProjectViewerComponent, resolve: { title: DynamicTitleResolve } },
 ];
 
 @NgModule({
@@ -33,8 +32,6 @@ const projectRoutes: Routes = [
     imports: [
         SharedModule,
         RouterModule.forChild(projectRoutes),
-
-        NgChartModule
     ]
 })
 export class ProjectModule { }