+++ AngularClient/src/main/shared/ng-chart/chart-data-item.ts
@@ -1,5 +1,0 @@
-﻿import { Subject } from "rxjs";
-
-export class ChartDataItem {
-    constructor(public name: string, public value: number, public valueUpdated: Subject<number>) { }
-}