+++ AngularClient/src/main/shared/shared.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { MomentModule } from "angular2-moment";
 
-import { NgChartModule } from "./ng-chart/ng-chart.module";
 import { MaterialModule } from "./material.module";
 
 @NgModule({
@@ -11,8 +10,7 @@ import { MaterialModule } from "./material.module";
         CommonModule,
         FormsModule,
         MaterialModule,
-        MomentModule,
-        NgChartModule
+        MomentModule
     ]
 })
 export class SharedModule { }