+++ AngularClient/src/main/user/user.module.ts
@@ -9,7 +9,7 @@ import { UserService } from "./user.service";
 import { DynamicTitleResolve } from "../core/dynamic-title-resolve.service";
 
 const userRoutes: Routes = [
-    { path: ":username", component: ProfileComponent, resolve: { title: DynamicTitleResolve } }
+    { path: "users/:username", component: ProfileComponent, resolve: { title: DynamicTitleResolve } }
 ];
 
 @NgModule({