+++ BusinessObjects/Entities/Project.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 
 namespace forCrowd.Backbone.BusinessObjects.Entities
 {
@@ -11,25 +10,13 @@ namespace forCrowd.Backbone.BusinessObjects.Entities
             ElementSet = new HashSet<Element>();
         }
 
-        private string _key = string.Empty;
-
         public int Id { get; set; }
 
-        [Index("UX_Project_UserId_Key", 1, IsUnique = true)]
         public int UserId { get; set; }
 
         [Required]
         [StringLength(50)]
         public string Name { get; set; }
-
-        [Required]
-        [StringLength(250)]
-        [Index("UX_Project_UserId_Key", 2, IsUnique = true)]
-        public string Key
-        {
-            get => _key;
-            set => _key = value.Replace(" ", "-");
-        }
 
         [Required]
         [StringLength(500)]